requirements.txt
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
xgboost==2.0.3
lightgbm==4.3.0
optuna==3.6.1
mlflow==2.15.1
pyyaml==6.0.2
matplotlib==3.9.0
ydata-profiling==4.8.3
joblib==1.4.2
tqdm==4.66.4
